Refactor search pages to remove search phase, split FEATURED_QUESTS by purpose
