+++ services/app/src/Constants.tsx
@@ -23,7 +23,79 @@ export const MULTIPLAYER_SETTINGS = {
   websocketSession: ((NODE_ENV === 'production') ? 'wss://' : 'ws://') + splitURL[splitURL.length - 1] + '/ws/multiplayer/v1/session',
 };
 
-export const FEATURED_QUESTS: Quest[] = [ // Featured quest ids generated from publishing, but don't leave them published!
+export const BUNDLED_QUESTS: Quest[] = [ // TODO - actually put GM quests here.
+  new Quest({
+    id: '0BzrQOdaJcH9MU3Z4YnE2Qi1oZGs',
+    partition: 'expedition-public',
+    theme: 'base',
+    official: true,
+    title: 'Oust Albanus',
+    summary: 'Your party encounters a smelly situation.',
+    author: 'Scott Martin',
+    publishedurl: 'quests/oust_albanus.xml',
+    minplayers: 1,
+    maxplayers: 6,
+    mintimeminutes: 20,
+    maxtimeminutes: 40,
+    genre: 'Comedy',
+    contentrating: 'Kid-friendly',
+    language: 'English',
+  }),
+  new Quest({
+    id: '0B7K9abSH1xEOUUR1Z0lncm9NRjQ',
+    partition: 'expedition-public',
+    theme: 'base',
+    official: true,
+    title: 'Dungeon Crawl',
+    summary: 'How deep can you delve?',
+    author: 'Todd Medema',
+    publishedurl: 'quests/dungeon_crawl.xml',
+    minplayers: 1,
+    maxplayers: 6,
+    mintimeminutes: 20,
+    maxtimeminutes: 60,
+    genre: 'Drama',
+    contentrating: 'Kid-friendly',
+    language: 'English',
+  }),
+  new Quest({
+    id: '0B7K9abSH1xEORjdkMWtTY3ZtNGs',
+    partition: 'expedition-public',
+    theme: 'base',
+    official: true,
+    title: 'Mistress Malaise',
+    summary: 'Mystery, Misfortune, and a Mistress.',
+    author: 'Scott Martin',
+    publishedurl: 'quests/mistress_malaise.xml',
+    minplayers: 1,
+    maxplayers: 6,
+    mintimeminutes: 30,
+    maxtimeminutes: 60,
+    genre: 'Drama',
+    contentrating: 'Kid-friendly',
+    language: 'English',
+  }),
+];
+export const GM_QUESTS: Quest[] = [ // TODO - actually put GM quests here.
+  new Quest({
+    id: '0BzrQOdaJcH9MU3Z4YnE2Qi1oZGs',
+    partition: 'expedition-public',
+    theme: 'base',
+    official: true,
+    title: 'Test GM Quest',
+    summary: 'This is totally a quest.',
+    author: 'Scott Martin',
+    publishedurl: 'quests/oust_albanus.xml',
+    minplayers: 1,
+    maxplayers: 6,
+    mintimeminutes: 20,
+    maxtimeminutes: 40,
+    genre: 'Comedy',
+    contentrating: 'Kid-friendly',
+    language: 'English',
+  }),
+];
+export const TUTORIAL_QUESTS: Quest[] = [ // Featured quest ids generated from publishing, but don't leave them published!
   new Quest({
     id: '0B7K9abSH1xEOeEZSVVMwNHNqaFE',
     partition: 'expedition-public',
@@ -78,61 +150,10 @@ export const FEATURED_QUESTS: Quest[] = [ // Featured quest ids generated from p
     contentrating: 'Kid-friendly',
     language: 'English',
   }),
-  new Quest({
-    id: '0BzrQOdaJcH9MU3Z4YnE2Qi1oZGs',
-    partition: 'expedition-public',
-    theme: 'base',
-    official: true,
-    title: 'Oust Albanus',
-    summary: 'Your party encounters a smelly situation.',
-    author: 'Scott Martin',
-    publishedurl: 'quests/oust_albanus.xml',
-    minplayers: 1,
-    maxplayers: 6,
-    mintimeminutes: 20,
-    maxtimeminutes: 40,
-    genre: 'Comedy',
-    contentrating: 'Kid-friendly',
-    language: 'English',
-  }),
-  new Quest({
-    id: '0B7K9abSH1xEORjdkMWtTY3ZtNGs',
-    partition: 'expedition-public',
-    theme: 'base',
-    official: true,
-    title: 'Mistress Malaise',
-    summary: 'Mystery, Misfortune, and a Mistress.',
-    author: 'Scott Martin',
-    publishedurl: 'quests/mistress_malaise.xml',
-    minplayers: 1,
-    maxplayers: 6,
-    mintimeminutes: 30,
-    maxtimeminutes: 60,
-    genre: 'Drama',
-    contentrating: 'Kid-friendly',
-    language: 'English',
-  }),
-  new Quest({
-    id: '0B7K9abSH1xEOUUR1Z0lncm9NRjQ',
-    partition: 'expedition-public',
-    theme: 'base',
-    official: true,
-    title: 'Dungeon Crawl',
-    summary: 'How deep can you delve?',
-    author: 'Todd Medema',
-    publishedurl: 'quests/dungeon_crawl.xml',
-    minplayers: 1,
-    maxplayers: 6,
-    mintimeminutes: 20,
-    maxtimeminutes: 60,
-    genre: 'Drama',
-    contentrating: 'Kid-friendly',
-    language: 'English',
-  }),
 ];
 
 if (NODE_ENV === 'dev') { // http://quests.expeditiongame.com/#0B7K9abSH1xEOV3M2bTVMdWc4NVk
-  FEATURED_QUESTS.unshift(new Quest({
+  TUTORIAL_QUESTS.unshift(new Quest({
     id: '1',
     partition: 'expedition-public',
     title: 'Test quest',
