+++ services/app/src/actions/Search.tsx
@@ -2,30 +2,29 @@ import {QuestSearchResponse} from 'api/Handlers';
 import * as Redux from 'redux';
 import {Quest} from 'shared/schema/Quests';
 import {openSnackbar} from '../actions/Snackbar';
-import {AUTH_SETTINGS, FEATURED_QUESTS} from '../Constants';
-import {CardPhase, ExpansionsType, SearchSettings, SettingsType} from '../reducers/StateTypes';
+import {AUTH_SETTINGS, TUTORIAL_QUESTS} from '../Constants';
+import {ExpansionsType, SearchParams, SettingsType} from '../reducers/StateTypes';
 import {SearchResponseAction} from './ActionTypes';
 import {remoteify} from './ActionTypes';
 import {toCard} from './Card';
 import {previewQuest} from './Quest';
 
 // TODO: Make search options propagate to other clients
-export const search = remoteify(function search(a: {search: SearchSettings, settings: SettingsType}, dispatch: Redux.Dispatch<any>) {
-  const params = {...a.search};
+export const search = remoteify(function search(a: {params: SearchParams, settings: SettingsType}, dispatch: Redux.Dispatch<any>) {
+  const params = {...a.params};
   Object.keys(params).forEach((key: string) => {
     if ((params as any)[key] === null) {
       delete (params as any)[key];
     }
   });
   params.players = a.settings.numPlayers;
-  const dispatchPhase = (params.partition === 'expedition-private') ? 'PRIVATE' : 'SEARCH';
   params.expansions = Object.keys(a.settings.contentSets).filter( (key) => a.settings.contentSets[key] ) as ExpansionsType[],
-  dispatch(toCard({name: 'SEARCH_CARD', phase: dispatchPhase as CardPhase}));
+  dispatch(toCard({name: 'SEARCH_CARD'}));
   dispatch(getSearchResults(params, (response: QuestSearchResponse) => {
     dispatch({
       quests: response.quests,
       error: response.error,
-      search: params,
+      params,
       type: 'SEARCH_RESPONSE',
     } as SearchResponseAction);
   }));
@@ -37,8 +36,8 @@ export const searchAndPlay = remoteify(function searchAndPlay(id: string, dispat
   const params = {
     id,
     partition: 'expedition-public',
-  } as SearchSettings;
-  const featuredQuest = FEATURED_QUESTS.filter((q) => q.id === id);
+  } as SearchParams;
+  const featuredQuest = TUTORIAL_QUESTS.filter((q) => q.id === id);
   if (featuredQuest.length === 1) {
     dispatch(previewQuest({quest: featuredQuest[0]}));
   } else {
@@ -46,7 +45,7 @@ export const searchAndPlay = remoteify(function searchAndPlay(id: string, dispat
       dispatch({
         quests: response.quests,
         error: response.error,
-        search: {}, // Don't specify search params because this one's weird and uses ID
+        params: {}, // Don't specify search params because this one's weird and uses ID
         type: 'SEARCH_RESPONSE',
       } as SearchResponseAction);
       if (response.quests.length === 0) {
@@ -63,7 +62,7 @@ export const searchAndPlay = remoteify(function searchAndPlay(id: string, dispat
 });
 
 // TODO switch to fetch since this never loads local files
-function getSearchResults(params: SearchSettings, callback: (response: QuestSearchResponse) => void) {
+function getSearchResults(params: SearchParams, callback: (response: QuestSearchResponse) => void) {
   return (dispatch: Redux.Dispatch<any>) => {
     // Clear previous results
     dispatch({type: 'SEARCH_REQUEST'});
