+++ services/app/src/components/views/FeaturedQuests.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import {Quest} from 'shared/schema/Quests';
-import {CardName, SettingsType, UserState} from '../../reducers/StateTypes';
+import {SettingsType, UserState} from '../../reducers/StateTypes';
 import Button from '../base/Button';
 import Card from '../base/Card';
 
@@ -11,8 +11,6 @@ export interface StateProps {
 }
 
 export interface DispatchProps {
-  toCard: (name: CardName) => any;
-  onSearchSelect: (user: UserState, settings: SettingsType) => any;
   onQuestSelect: (quest: Quest) => any;
 }
 
@@ -35,46 +33,9 @@ const Tutorials = (props: Props): JSX.Element => {
       );
     });
 
-  /*
-  import {openWindow} from '../../Globals';
-  {!props.settings.simulator && <Button onClick={() => props.onSearchSelect(props.user, props.settings)} id="morequests">
-        <div className="questButtonWithIcon">
-          <div className="title"><img className="inline_icon" src="images/book_small.svg"/>More Quests</div>
-          <div className="summary">Explore and play community-written quests.</div>
-        </div>
-      </Button>
-      }
-
-      {!props.settings.simulator && props.settings.experimental &&
-        <Button onClick={() => props.toCard('SAVED_QUESTS')} id="saved">
-        <div className="questButtonWithIcon">
-          <div className="title"><img className="inline_icon" src="images/compass_small.svg"/>Saved & Offline Quests</div>
-        </div>
-      </Button>
-      }
-      {!props.settings.simulator &&
-        <Button onClick={() => props.toCard('QUEST_HISTORY')} id="history">
-        <div className="questButtonWithIcon">
-          <div className="title"><img className="inline_icon" src="images/compass_small.svg"/>Quest History</div>
-        </div>
-      </Button>
-      }
-      {!props.settings.simulator && <Button onClick={() => props.toCard('ADVANCED')} id="tools">
-        <div className="questButtonWithIcon">
-          <div className="title"><img className="inline_icon" src="images/roll_small.svg"/>Tools</div>
-        </div>
-      </Button>}
-      <Button onClick={() => openWindow('https://expeditiongame.com/store?utm_source=app')}>
-        <div className="questButtonWithIcon">
-          <div className="title"><img className="inline_icon" src="images/loot_small.svg"/>Shop</div>
-        </div>
-      </Button>
-  */
-
   return (
-    <Card title="Tutorial Quests" icon="helper">
+    <Card title="Tutorial Quests" icon="helper" showReturn={false}>
       {items}
-
     </Card>
   );
 };
