+++ services/app/src/components/views/FeaturedQuestsContainer.tsx
@@ -2,17 +2,14 @@ import {connect} from 'react-redux';
 import Redux from 'redux';
 
 import {Quest} from 'shared/schema/Quests';
-import {toCard} from '../../actions/Card';
 import {previewQuest} from '../../actions/Quest';
-import {search} from '../../actions/Search';
-import {FEATURED_QUESTS} from '../../Constants';
-import {initialSearch} from '../../reducers/Search';
-import {AppState, CardName, SettingsType, UserState} from '../../reducers/StateTypes';
+import {TUTORIAL_QUESTS} from '../../Constants';
+import {AppState} from '../../reducers/StateTypes';
 import FeaturedQuests, {DispatchProps, StateProps} from './FeaturedQuests';
 
 const mapStateToProps = (state: AppState): StateProps => {
   return {
-    quests: FEATURED_QUESTS,
+    quests: TUTORIAL_QUESTS,
     settings: state.settings,
     user: state.user,
   };
@@ -20,19 +17,6 @@ const mapStateToProps = (state: AppState): StateProps => {
 
 const mapDispatchToProps = (dispatch: Redux.Dispatch<any>): DispatchProps => {
   return {
-    toCard(name: CardName): void {
-      dispatch(toCard({name}));
-    },
-    onSearchSelect(user: UserState, settings: SettingsType): void {
-      if (user && user.loggedIn) {
-        dispatch(search({
-          search: initialSearch.search,
-          settings,
-        }));
-      } else {
-        dispatch(toCard({name: 'SEARCH_CARD', phase: 'DISCLAIMER'}));
-      }
-    },
     onQuestSelect(quest: Quest): void {
       dispatch(previewQuest({quest}));
     },
