+++ services/app/src/components/views/QuestPreview.test.tsx
@@ -1,7 +1,7 @@
 import {configure, shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import {Quest} from 'shared/schema/Quests';
-import {FEATURED_QUESTS} from '../../Constants';
+import {TUTORIAL_QUESTS} from '../../Constants';
 import {initialSettings} from '../../reducers/Settings';
 import QuestPreview, {Props} from './QuestPreview';
 configure({ adapter: new Adapter() });
@@ -19,7 +19,7 @@ describe('QuestPreview', () => {
       isDirectLinked: false,
       savedInstances: [],
       lastPlayed: null,
-      quest: new Quest({...FEATURED_QUESTS.filter((el) => el.title === questTitle)[0], ...questOverrides}),
+      quest: new Quest({...TUTORIAL_QUESTS.filter((el) => el.title === questTitle)[0], ...questOverrides}),
       settings: {...initialSettings, experimental: true},
       onPlay: jasmine.createSpy('onPlay'),
       onPlaySaved: jasmine.createSpy('onPlaySaved'),
@@ -37,7 +37,7 @@ describe('QuestPreview', () => {
   }
 
   test('renders selected quest details', () => {
-    const quest = FEATURED_QUESTS.filter((el) => el.title === 'Learning to Adventure')[0];
+    const quest = TUTORIAL_QUESTS.filter((el) => el.title === 'Learning to Adventure')[0];
     const {wrapper} = setup(quest.title);
     expect(wrapper.render().html()).toContain(quest.title);
     expect(wrapper.render().html()).toContain(quest.genre);
@@ -83,7 +83,7 @@ describe('QuestPreview', () => {
   });
 
   test('allows save for offline play', () => {
-    const quest = new Quest({...FEATURED_QUESTS[0], publishedurl: 'http://somenonlocalurl'});
+    const quest = new Quest({...TUTORIAL_QUESTS[0], publishedurl: 'http://somenonlocalurl'});
     const {props, wrapper} = setup({quest});
     wrapper.find('#offlinesave').simulate('click');
     expect(props.onSave).toHaveBeenCalledWith(quest);
@@ -108,7 +108,7 @@ describe('QuestPreview', () => {
   });
 
   test('disallows saving local quests for offline play', () => {
-    const quest = new Quest({...FEATURED_QUESTS[0], publishedurl: 'quests/localquest.xml'});
+    const quest = new Quest({...TUTORIAL_QUESTS[0], publishedurl: 'quests/localquest.xml'});
     const {props, wrapper} = setup({quest});
     expect(wrapper.find('#offlinesave').exists()).toEqual(false);
   });
@@ -121,7 +121,7 @@ describe('QuestPreview', () => {
   });
 
   test('indicates that horror and future expansions are required', () => {
-    const quest = FEATURED_QUESTS.filter((el) => el.title === 'Learning 3: The Future')[0];
+    const quest = TUTORIAL_QUESTS.filter((el) => el.title === 'Learning 3: The Future')[0];
     const {wrapper} = setup({quest});
     expect(wrapper.html()).toContain('horror');
     expect(wrapper.html()).toContain('future');
