+++ services/app/src/components/views/Search.tsx
@@ -1,28 +1,20 @@
-import FormControl from '@material-ui/core/FormControl';
-import InputLabel from '@material-ui/core/InputLabel';
-import NativeSelect from '@material-ui/core/NativeSelect';
-import TextField from '@material-ui/core/TextField';
 import DoneIcon from '@material-ui/icons/Done';
 import OfflinePin from '@material-ui/icons/OfflinePin';
 import StarsIcon from '@material-ui/icons/Stars';
 import * as React from 'react';
 import Truncate from 'react-truncate';
-import {CONTENT_RATING_DESC, GenreType, LANGUAGES} from 'shared/schema/Constants';
 import {Quest} from 'shared/schema/Quests';
-import {PLAYTIME_MINUTES_BUCKETS} from '../../Constants';
 import {formatPlayPeriod, smartTruncateSummary} from '../../Format';
-import {SearchPhase, SearchSettings, SearchState, SettingsType, UserQuestHistory, UserState} from '../../reducers/StateTypes';
+import {SearchParams, SearchState, SettingsType, UserQuestHistory, UserState} from '../../reducers/StateTypes';
 import Button from '../base/Button';
 import Card from '../base/Card';
-import Checkbox from '../base/Checkbox';
 import StarRating from '../base/StarRating';
 
 const Moment = require('moment');
 
 export interface StateProps extends SearchState {
   isDirectLinked: boolean;
-  phase: SearchPhase;
-  search: SearchSettings;
+  params: SearchParams;
   settings: SettingsType;
   user: UserState;
   questHistory: UserQuestHistory;
@@ -34,187 +26,9 @@ export interface DispatchProps {
   onLoginRequest: (subscribe: boolean) => void;
   onQuest: (quest: Quest) => void;
   onReturn: () => void;
-  onSearch: (search: SearchSettings, settings: SettingsType) => void;
 }
 
 export interface Props extends StateProps, DispatchProps {}
-
-// We make this a react component to hold a bit of state and avoid sending
-// redux actions for every single change to input.
-export interface SearchSettingsCardProps {
-  onSearch: (search: SearchSettings, settings: SettingsType) => void;
-  search: SearchSettings;
-  settings: SettingsType;
-  user: UserState;
-}
-
-export class SearchSettingsCard extends React.Component<SearchSettingsCardProps, {}> {
-  public state: SearchSettings;
-
-  constructor(props: SearchSettingsCardProps) {
-    super(props);
-    this.state = this.props.search;
-  }
-
-  public onChange(attrib: string, value: any) {
-    this.setState({[attrib]: value});
-  }
-
-  public submit(e: React.FormEvent | React.MouseEvent | undefined) {
-    if (e) {
-      e.preventDefault();
-    }
-    this.props.onSearch(this.state, this.props.settings);
-  }
-
-// TODO remove the clutter here / move to Theme.tsx
-  public render() {
-    const rating = (this.state.contentrating) ? CONTENT_RATING_DESC[this.state.contentrating] : undefined;
-    const timeBuckets = PLAYTIME_MINUTES_BUCKETS.map((minutes: number, index: number) => {
-      return <option key={index} value={minutes}>{`${minutes} min`}</option>;
-    });
-    // TODO Once we have 3 romance quests, change code to just display genre list
-    const visibleGenres: GenreType[] = ['Comedy', 'Drama', 'Horror', 'Mystery'];
-    return (
-      <Card title="Quest Search">
-        <form className="searchForm" autoComplete="off" onSubmit={(e: React.FormEvent) => {this.submit(e); }}>
-          <div className="searchDescription">
-            For {this.props.settings.numPlayers} adventurer{this.props.settings.numPlayers > 1 ? 's' : ''} with {this.props.settings.contentSets.horror ? 'The Horror' : 'the base game'} (based on settings)
-          </div>
-          <FormControl fullWidth={true}>
-            <TextField
-              id="text"
-              className="textfield"
-              fullWidth={true}
-              label="text search - title, author, ID"
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => this.onChange('text', e.target.value)}
-              onFocus={(e: any) => e.target.scrollIntoView()}
-              value={this.state.text}
-            />
-          </FormControl>
-          <FormControl className="selectfield" fullWidth={true}>
-            <InputLabel htmlFor="order">Sort by</InputLabel>
-            <NativeSelect
-              inputProps={{
-                id: 'order',
-              }}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>, c: React.ReactNode) => this.onChange('order', e.target.value)}
-              value={this.state.order}
-            >
-              <option value="+ratingavg">Highest rated</option>
-              <option value="-created">Newest</option>
-              <option value="+title">Title (A-Z)</option>
-              <option value="-title">Title (Z-A)</option>
-            </NativeSelect>
-          </FormControl>
-          <FormControl className="selectfield halfLeft ranged">
-            <InputLabel htmlFor="mintimeminutes">Minimum time</InputLabel>
-            <NativeSelect
-              inputProps={{
-                id: 'mintimeminutes',
-              }}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>, c: React.ReactNode) => this.onChange('mintimeminutes', e.target.value)}
-              value={this.state.mintimeminutes}
-            >
-              <option value={undefined}>Any length</option>
-              {timeBuckets}
-            </NativeSelect>
-          </FormControl>
-          <FormControl className="selectfield halfRight">
-            <InputLabel htmlFor="maxtimeminutes">Maximum time</InputLabel>
-            <NativeSelect
-              inputProps={{
-                id: 'maxtimeminutes',
-              }}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>, c: React.ReactNode) => this.onChange('maxtimeminutes', e.target.value)}
-              value={this.state.maxtimeminutes}
-            >
-              <option value={undefined}>Any length</option>
-              {timeBuckets}
-            </NativeSelect>
-          </FormControl>
-          <FormControl className="selectfield halfLeft">
-            <InputLabel htmlFor="age">Recency</InputLabel>
-            <NativeSelect
-              inputProps={{
-                id: 'age',
-              }}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>, c: React.ReactNode) => this.onChange('age', e.target.value)}
-              value={this.state.age}
-            >
-              <option value={undefined}>All time</option>
-              <option value={31536000}>Published this year</option>
-              <option value={2592000}>Published this month</option>
-              <option value={604800}>Published this week</option>
-            </NativeSelect>
-          </FormControl>
-          <FormControl className="selectfield halfRight">
-            <InputLabel htmlFor="language">Language</InputLabel>
-            <NativeSelect
-              inputProps={{
-                id: 'language',
-              }}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>, c: React.ReactNode) => this.onChange('language', e.target.value)}
-              value={this.state.language}
-            >
-              {LANGUAGES.map((language: string, i: number) => <option key={i} value={language}>{language}</option>)}
-            </NativeSelect>
-          </FormControl>
-          <FormControl className="selectfield halfLeft">
-            <InputLabel htmlFor="genre">Genre</InputLabel>
-            <NativeSelect
-              inputProps={{
-                id: 'genre',
-              }}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>, c: React.ReactNode) => this.onChange('genre', e.target.value)}
-              value={this.state.genre}
-            >
-              <option value={undefined}>All genres</option>
-              {visibleGenres.map((genre: string, i: number) => <option key={i} value={genre}>{genre}</option>)}
-            </NativeSelect>
-          </FormControl>
-          <FormControl className="selectfield halfRight">
-            <InputLabel htmlFor="contentrating">Content Rating</InputLabel>
-            <NativeSelect
-              inputProps={{
-                id: 'contentrating',
-              }}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>, c: React.ReactNode) => this.onChange('contentrating', e.target.value)}
-              value={this.state.contentrating}
-            >
-              <option value={undefined}>All ratings</option>
-              <option value="Kid-friendly">Kid-friendly</option>
-              <option value="Teen">Teen</option>
-              <option value="Adult">Adult</option>
-            </NativeSelect>
-          </FormControl>
-          <FormControl className="selectfield halfLeft">
-            <InputLabel htmlFor="requirespenpaper">Requires Pen & Paper</InputLabel>
-            <NativeSelect
-              inputProps={{
-                id: 'requirespenpaper',
-              }}
-              onChange={(e: React.ChangeEvent<HTMLSelectElement>, c: React.ReactNode) => this.onChange('requirespenpaper', e.target.value)}
-              value={this.state.requirespenpaper}
-            >
-              <option value={undefined}>No Preference</option>
-              <option value="true">Yes</option>
-              <option value="false">No</option>
-            </NativeSelect>
-          </FormControl>
-          {rating && <div className="ratingDescription">
-            <span>"{this.state.contentrating}" rating means: {rating.summary}</span>
-          </div>}
-          <Button onClick={(e: React.FormEvent) => {this.submit(e); }} id="search">Search</Button>
-        </form>
-      </Card>
-    );
-  }
-}
-
-function renderSettings(props: Props): JSX.Element {
-  return (<SearchSettingsCard search={props.search} settings={props.settings} onSearch={props.onSearch} user={props.user} />);
-}
 
 export interface SearchResultProps {
   index: number;
@@ -222,12 +36,12 @@ export interface SearchResultProps {
   lastPlayed: Date | null;
   onQuest: (quest: Quest) => void;
   quest: Quest;
-  search: SearchSettings;
+  params: SearchParams;
   offlineQuests: {[id: string]: boolean};
 }
 
 export function renderResult(props: SearchResultProps): JSX.Element {
-  const orderField = props.search.order && props.search.order.substring(1);
+  const orderField = props.params.order && props.params.order.substring(1);
   const quest = props.quest;
   let orderDetails = <span></span>;
   if (orderField) {
@@ -287,17 +101,46 @@ export function renderResult(props: SearchResultProps): JSX.Element {
   );
 }
 
-export function renderResults(props: Props, hideHeader?: boolean): JSX.Element {
+export function search(props: Props): JSX.Element {
+
+  // TODO Use BUNDLED_QUESTS when user not logged in
+
+  // TODO Private quests
+  /*
+  <Button id="selectPrivateQuests" onClick={() => props.onPrivateQuestsSelect(props.settings, props.user)}>
+    <div className="questButtonWithIcon">
+      <div className="title">Private Quests</div>
+      <div className="summary">View quests you've published privately with the Quest Creator (uses your current player count!)</div>
+    </div>
+  </Button>
+  dispatch(ensureLogin())
+        .then((u: UserState) => {
+          return dispatch(search({
+            params: {
+              order: '-published',
+              owner: u.id,
+              partition: 'expedition-private',
+            },
+            settings: {
+              ...settings,
+              contentSets: {
+                horror: true,
+              },
+            },
+          }));
+        });
+  */
+
   let content: JSX.Element | JSX.Element[];
   const numResults = (props.results || []).length;
   if (numResults === 0 && !props.searching) {
     content = (
       <div className="searchDescription">
         <h2>No quests found</h2>
-        {!hideHeader && <span>
+        <span>
           <p>Try broadening the search by using fewer filters.</p>
           <p>If you still see no results, file feedback from the top corner menu.</p>
-        </span>}
+        </span>
         <Button className="filter_button" onClick={() => props.onFilter()} id="filter">Modify Search</Button>
       </div>
     );
@@ -315,7 +158,7 @@ export function renderResults(props: Props, hideHeader?: boolean): JSX.Element {
       return renderResult({
         index,
         quest,
-        search: props.search,
+        params: props.params,
         onQuest: props.onQuest,
         lastLogin: props.user.lastLogin,
         lastPlayed: (props.questHistory.list[quest.id] || {}).lastPlayed,
@@ -328,7 +171,7 @@ export function renderResults(props: Props, hideHeader?: boolean): JSX.Element {
     <Card
       title="Quest Search Results"
       className="search_card"
-      header={(hideHeader) ? undefined : <div className="searchHeader">
+      header={<div className="searchHeader">
         <Button className="searchResultInfo" disabled={true}>{props.results.length} quests for {props.settings.numPlayers} <img className="inline_icon" src="images/adventurer_small.svg"/></Button>
         <Button className="filter_button" onClick={() => props.onFilter()} id="filter">Filter &amp; Sort ></Button>
       </div>}
@@ -338,67 +181,4 @@ export function renderResults(props: Props, hideHeader?: boolean): JSX.Element {
   );
 }
 
-// We make this a react component to hold a bit of state and avoid sending
-// redux actions for every single change to input.
-interface SearchDisclaimerCardProps {
-  onLoginRequest: (subscribe: boolean) => void;
-}
-
-class SearchDisclaimerCard extends React.Component<SearchDisclaimerCardProps, {}> {
-  public state: {
-    subscribe: boolean;
-  };
-
-  constructor(props: SearchDisclaimerCardProps) {
-    super(props);
-    this.state = {
-      subscribe: false,
-    };
-  }
-
-  public onSubscribeChange(value: boolean) {
-    this.setState({subscribe: value});
-  }
-
-  public render() {
-    return (
-      <Card title="Disclaimer">
-        <p>
-          Community quests are written by adventurers like yourselves using the free quest creator (Quests.ExpeditionGame.com).
-          We offer no guarantees for the quests you are about to play, but do our best to review them for quality, and provide players with
-          the ability to rate, review and report quests.
-        </p>
-        <p>
-          We use your Google email as your identity when rating and reviewing quests.
-        </p>
-        <p>
-          You must log in to continue:
-        </p>
-        <Checkbox label="Join the Mailing List" value={this.state.subscribe} onChange={(v: boolean) => { this.onSubscribeChange(v); }}>
-          Learn about the latest quests, features and more - once per month!
-        </Checkbox>
-        <Button onClick={(e) => this.props.onLoginRequest(this.state.subscribe)}>Continue with Google</Button>
-      </Card>
-    );
-  }
-}
-function renderDisclaimer(props: Props): JSX.Element {
-  return (<SearchDisclaimerCard onLoginRequest={props.onLoginRequest}/>);
-}
-
-const Search = (props: Props): JSX.Element => {
-  switch (props.phase) {
-    case 'DISCLAIMER':
-      return renderDisclaimer(props);
-    case 'SETTINGS':
-      return renderSettings(props);
-    case 'SEARCH':
-      return renderResults(props);
-    case 'PRIVATE':
-      return renderResults(props, true);
-    default:
-      throw new Error('Unknown search phase ' + props.phase);
-  }
-};
-
-export default Search;
+export default search;
